Refactor users.admin
